+++ src/compareImage.js
@@ -2,20 +2,23 @@
 import Jimp from 'jimp';
 import logger from './logger';
 
-const compareImage = (referencePath, testPath, misMatchThreshold, testName) =>
+const compareImage = (options, testName) =>
   new Promise((resolve, reject) => {
-    const referenceFile = `${referencePath}/${testName}.png`;
-    const testFile = `${testPath}/${testName}.png`;
+    const referenceFile = `${options.screenshots}/${testName}.png`;
+    const testFile = `${options.testReportPath}/${testName}.png`;
     Jimp.read(referenceFile).then((referenceImage) => {
       Jimp.read(testFile).then((testImage) => {
         const distance = Jimp.distance(referenceImage, testImage);
-        const diff = Jimp.diff(referenceImage, testImage, misMatchThreshold);
-        logger.log('Writting the diff image to disk');
-        if (distance < misMatchThreshold || diff.percent < misMatchThreshold) {
+        const diff = Jimp.diff(referenceImage, testImage, options.misMatchThreshold);
+        if (distance < options.misMatchThreshold || diff.percent < options.misMatchThreshold) {
           return resolve('No mismatch found!');
         }
-        diff.image.write(`${testPath}/${testName}_differencified.png`);
-        logger.error(`Result ->  distance:${distance} diff:${diff.percent} misMatchThreshold:${misMatchThreshold}`);
+        if (options.saveDifferencifiedImage) {
+          logger.log('Writting the diff image to disk');
+          diff.image.write(`${options.testPath}/${testName}_differencified.png`);
+        }
+        logger.error(`Result ->  distance:${distance} diff:${diff.percent} 
+          misMatchThreshold:${options.misMatchThreshold}`);
         return reject('Mismatch found!');
       }).catch((err) => {
         logger.error(err);