+++ src/compareImage.test.js
@@ -1,10 +1,11 @@
 import Jimp from 'jimp';
 import compareImage from './compareImage';
 import logger from './logger';
+import { globalConfig } from './defaultConfig';
 
 Jimp.read = path =>
   new Promise((resolve, reject) => {
-    if (path === 'path/test1.png') {
+    if (path === './screenshots/test1.png' || path === './differencify_report/test1.png') {
       return resolve('image');
     }
     return reject('error');
@@ -29,7 +30,7 @@ describe('Compare Image', () => {
     loggerCalls = [];
   });
   it('returns correct values', async () =>
-    compareImage('path', 'path', 0.1, 'test1')
+    compareImage(globalConfig, 'test1')
     .catch((result) => {
       expect(result).toEqual('Failed to read test image!');
       expect(loggerCalls[0]).toEqual('Writting the diff image to disk');