+++ src/defaultConfig.js
@@ -1,6 +1,7 @@
 const globalConfig = {
   screenshots: './screenshots',
   testReportPath: './differencify_report',
+  saveDifferencifiedImage: true,
   debug: false,
   visible: false,
   timeout: 30000,